added GitHub connector to the build (converted to the new architecture)

+++ common/src/main/scala/com/taskadapter/model/GProject.scala
@@ -3,7 +3,7 @@ package com.taskadapter.model
 /**
   * Generic "Project" object. Internal representation for "projects".
   */
-case class GProject(id: Integer, name: String, key: String, description: String, homepage: String)
+case class GProject(id: Long, name: String, key: String, description: String, homepage: String)
   extends NamedKeyedObject {
   override def getName: String = name
